перенесена рабочая версия MainPage
