+++ src/pages/NotFound/NotFound.jsx
@@ -1,0 +1,5 @@
+export const NotFound = () => {
+    return (
+        <h1>404</h1>
+    )
+}
