+++ src/router/routes.js
@@ -1,0 +1,6 @@
+export const routes = {
+    main: "/",
+    login: "/login",
+    register: "/register",
+    notFound: "*"   
+}
